Refactor the component names in test files
